Create Android APK version of the agent (#2019)
* Add android flag to build

Added gomobile

* [Android] agent working as app.

Intermediate checkin, not complete.  Significant items
[] in this branch, logging is effectively commented out (even for non
android builds).  Need sustainable approach
[] starting of api server commented out (for all platforms)  Need
 sustainable approach
[] runs as app, but stops when app is stopped. checkpoint build as
 transition to service

* First move of sample app (doesn't even compile yet)

* Basic app working with gobind
Uses Java top level app (which allows us to have service intent)

Runs as service even when app stopped.

Still needs logging fixed

* Add launcher icon (looks pretty)
Remove out-of-place AndroidManifest.xml; don't need it and it's
confusing

* Move default configuration to /data/datadog-agent
Location is configurable in included android.yaml; customer must break
open APK, modify android.yaml, and re-sign to use.

* Add ability to make datadog agent as aar, then link with app

* add gradle files

Add gradlew (build script for linux)

Add android CI job

remove other build targets for now...

Set default for bin_name to be android=false.
Re-add android apk target

* Add android:singleUser="true" /> at customer's request

* Add android tasks to invoke

* Fix invoke tasks for android.  Change gitlab to use those tasks

* Add android flag to build

Added gomobile

* [Android] agent working as app.

Intermediate checkin, not complete.  Significant items
[] in this branch, logging is effectively commented out (even for non
android builds).  Need sustainable approach
[] starting of api server commented out (for all platforms)  Need
 sustainable approach
[] runs as app, but stops when app is stopped. checkpoint build as
 transition to service

* First move of sample app (doesn't even compile yet)

* Basic app working with gobind
Uses Java top level app (which allows us to have service intent)

Runs as service even when app stopped.

Still needs logging fixed

* Add launcher icon (looks pretty)
Remove out-of-place AndroidManifest.xml; don't need it and it's
confusing

* Move default configuration to /data/datadog-agent
Location is configurable in included android.yaml; customer must break
open APK, modify android.yaml, and re-sign to use.

* Add ability to make datadog agent as aar, then link with app

* add gradle files

Add gradlew (build script for linux)

Add android CI job

remove other build targets for now...

Set default for bin_name to be android=false.
Re-add android apk target

* Add android:singleUser="true" /> at customer's request

* Add android tasks to invoke

* Fix invoke tasks for android.  Change gitlab to use those tasks

* Remove processing of assets (as it currently causes service to crash)

* Review feedback

+++ tasks/go.py
@@ -180,7 +180,7 @@ def misspell(ctx, targets):
         print("misspell found no issues")
 
 @task
-def deps(ctx, no_checks=False, core_dir=None, verbose=False):
+def deps(ctx, no_checks=False, core_dir=None, verbose=False, android=False):
     """
     Setup Go dependencies
     """
@@ -204,7 +204,7 @@ def deps(ctx, no_checks=False, core_dir=None, verbose=False):
     ctx.run("dep ensure{}".format(verbosity))
     # make sure PSUTIL is gone on windows; the dep ensure above will vendor it
     # in because it's necessary on other platforms
-    if sys.platform == 'win32':
+    if not android and sys.platform == 'win32':
         print("Removing PSUTIL on Windows")
         ctx.run("rd /s/q vendor\\github.com\\shirou\\gopsutil")
